implementing SSO in VOC

+++ app/controllers/user_sessions_controller.rb
@@ -8,41 +8,60 @@ class UserSessionsController < ApplicationController
 
 	# GET    /user_sessions/new(.:format)
   def new
-    @user_session = UserSession.new
-  end
-
-  # POST   /user_sessions(.:format)
-  def create
-    @user_session = UserSession.new(params[:user_session])
-    if @user_session.save
-      redirect_to root_path
-    else
-      render :new
+    token_good, user_info = authenticate_token(request, nil, true) 
+    if token_good == true
+      user = User.find_by_hhs_id(sso_user_hhs_id)
+      if user
+        UserSession.create(user,true)
+        redirect_back_or_default surveys_path              
+      else
+        render :unauthorized
+      end
     end
   end
+
+  # def new
+  #   @user_session = UserSession.new
+  # end
+  
+  # # POST   /user_sessions(.:format)
+  # def create
+  #   @user_session = UserSession.new(params[:user_session])
+  #   if @user_session.save
+  #     redirect_to root_path
+  #   else
+  #     render :new
+  #   end
+  # end
 
   # DELETE /user_sessions/:id(.:format)
   def destroy
-    current_user_session.destroy
-    flash[:notice] = "Logout successful!"
-    redirect_to login_url
+    flash[:notice] = "Logout successful"
+    current_user_session.destroy if current_user_session
+    logout(true)
   end
+
+  # def destroy
+  #   current_user_session.destroy
+  #   flash[:notice] = "Logout successful!"
+  #   redirect_to login_url
+  # end
 
   # POST   /user_sessions/do_pw_reset(.:format)
-  def do_pw_reset
-    user = User.find_by_email(params["email_address"])
-    if user
-      password = PasswordGenerator::generate_password(2,2,2,2)
-      user.password = password
-      user.password_confirmation = password
+  # def do_pw_reset
+  #   user = User.find_by_email(params["email_address"])
+  #   if user
+  #     password = PasswordGenerator::generate_password(2,2,2,2)
+  #     user.password = password
+  #     user.password_confirmation = password
 
-      if user.save_without_session_maintenance
-        UserSessionsMailer.reset_password(user, password).deliver
-      end
-    end
-    flash.notice = "An email has been sent to #{params["email_address"]}, if the account existed."
-    redirect_to login_path
-  end
+  #     if user.save_without_session_maintenance
+  #       UserSessionsMailer.reset_password(user, password).deliver
+  #     end
+  #   end
+  #   flash.notice = "An email has been sent to #{params["email_address"]}, if the account existed."
+  #   redirect_to login_path
+  # end
 
   private
